+++ Server/src/main/java/com/course/server/endpoint/ObjectServlet.java
@@ -3,6 +3,7 @@ package com.course.server.endpoint;
 import com.course.server.ApplicationServiceProvider;
 import com.course.server.database.Database;
 import com.course.server.domain.InventoryObject;
+import com.course.server.domain.Role;
 import com.course.server.service.JsonStream;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
@@ -12,6 +13,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.util.UUID;
 
 public class ObjectServlet extends HttpServlet
 {
@@ -25,43 +27,66 @@ public class ObjectServlet extends HttpServlet
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
     {
-        String inventoryNumber = req.getParameter("id");
-        InventoryObject object = applicationServiceProvider.database.getObject(inventoryNumber);
+        UUID issuedById = UUID.fromString(req.getParameter("by"));
 
-        resp.setContentType("text/json");
-        PrintWriter printWriter = resp.getWriter();
-        JsonStream stream = new JsonStream(printWriter);
-        stream.write(object);
-        printWriter.close();
+        if (applicationServiceProvider.authenticator.isValidUser(issuedById))
+        {
+            String inventoryNumber = req.getParameter("id");
+            InventoryObject object = applicationServiceProvider.database.getObject(inventoryNumber);
+
+            resp.setContentType("text/json");
+            PrintWriter printWriter = resp.getWriter();
+            JsonStream stream = new JsonStream(printWriter);
+            stream.write(object);
+            printWriter.close();
+        }
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
     {
-        BufferedReader reader = req.getReader();
-        JsonStream stream = new JsonStream(reader);
-        InventoryObject object = stream.readObject();
-        applicationServiceProvider.database.createObject(object);
+        UUID issuedById = UUID.fromString(req.getParameter("by"));
+
+        if (applicationServiceProvider.authenticator.isValidUser(issuedById) &&
+                applicationServiceProvider.database.getUser(issuedById).getRole().compareTo(Role.INVENTORY_OFFICER) >= 0)
+        {
+            BufferedReader reader = req.getReader();
+            JsonStream stream = new JsonStream(reader);
+            InventoryObject object = stream.readObject();
+            applicationServiceProvider.database.createObject(object);
+        }
     }
 
     @Override
     protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
     {
-        String inventoryNumber = req.getParameter("id");
-        BufferedReader reader = req.getReader();
-        JsonStream stream = new JsonStream(reader);
-        InventoryObject object = stream.readObject();
+        UUID issuedById = UUID.fromString(req.getParameter("by"));
 
-        if (inventoryNumber.equals(object.getInventoryNumber()))
+        if (applicationServiceProvider.authenticator.isValidUser(issuedById) &&
+                applicationServiceProvider.database.getUser(issuedById).getRole().compareTo(Role.INVENTORY_OFFICER) >= 0)
         {
-            applicationServiceProvider.database.updateObject(object);
+            String inventoryNumber = req.getParameter("id");
+            BufferedReader reader = req.getReader();
+            JsonStream stream = new JsonStream(reader);
+            InventoryObject object = stream.readObject();
+
+            if (inventoryNumber.equals(object.getInventoryNumber()))
+            {
+                applicationServiceProvider.database.updateObject(object);
+            }
         }
     }
 
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
     {
-        String inventoryNumber = req.getParameter("id");
-        applicationServiceProvider.database.deleteObject(inventoryNumber);
+        UUID issuedById = UUID.fromString(req.getParameter("by"));
+
+        if (applicationServiceProvider.authenticator.isValidUser(issuedById) &&
+                applicationServiceProvider.database.getUser(issuedById).getRole().compareTo(Role.ADMINISTRATOR) >= 0)
+        {
+            String inventoryNumber = req.getParameter("id");
+            applicationServiceProvider.database.deleteObject(inventoryNumber);
+        }
     }
 }
